refactor: refactor-pages
