+++ src/pages/create/page.tsx
@@ -1,0 +1,6 @@
+import React from "react"
+import exp from "constants"
+
+const Create = () => <h1>create</h1>
+
+export default Create
