+++ src/pages/login/pages.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import exp from "constants"
 
-const Edit = () => <h1>edit</h1>
+const Login = () => <h1>login</h1>
 
-export default Edit
+export default Login
