--- src/pages/play.tsx
@@ -1,6 +1,0 @@
-import React from "react"
-import exp from "constants"
-
-const Play = () => <h1>Play</h1>
-
-export default Play
