--- src/pages/top.tsx
@@ -1,6 +1,0 @@
-import React from "react"
-import exp from "constants"
-
-const Top = () => <h1>hello</h1>
-
-export default Top
